Update
All of the work done since the last push, including an updated spAnalyze
code to analyze the absorption features including code to separate the
iron triplets. Also, conv06X is a new script to convolve the SN2006X
data to the resolution and pixel scale of the other data to determine if
any variation would be seen.